auditor/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# concatenate_csvs.py: 10
click == 6.7

# concatenate_csvs.py: 9
# modules/utils.py: 10
pandas

# modules/dev_tools.py: 6
requests == 2.9.1

# auditor.py: 11
# modules/dev_tools.py: 7
websocket_client == 0.56.0
